Moved logger to LogHelper
