+++ extension/src/Logging/LogHelper.ts
@@ -1,5 +1,12 @@
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 const timestamp = require("time-stamp");
+import { NullLogger } from "./NullLogger";
+
+export let logger: ILogger = new NullLogger();
+
+export function setLogger(newLogger: ILogger): void {
+  logger = newLogger;
+}
 
 export interface ILogger {
   log(message?: string, ...optionalParams: string[]): void;
