--- extension/src/Logging/Logger.ts
@@ -1,8 +1,0 @@
-import { ILogger } from "./LogHelper";
-import { NullLogger } from "./NullLogger";
-
-export let logger: ILogger = new NullLogger();
-
-export function setLogger(newLogger: ILogger): void {
-  logger = newLogger;
-}
