+++ extension/src/PowerShell.ts
@@ -1,6 +1,6 @@
 import { ILogger } from "./Logging/LogHelper";
 import * as Shell from "node-powershell";
-import { logger } from "./Logging/Logger";
+import { logger } from "./Logging/LogHelper";
 
 export class Powershell {
   private startTime: Date = new Date();
